Working with animations is easier now. Shotting animation implamented.
Modified setter for an animation name in the player class, so you don't need to provide a string with weapon name and a animation type, you can just provide an animation type, everything else will be setup for you. Added ability to lock animations, that means that the animation won't be able to change for a certain number of frame, which you can set using Entity.animationLock.